wip: rename to text
