+++ packages/clippy-components/src/clippy-token-sample-text/README.md
@@ -1,16 +1,16 @@
-# `<clippy-token-sample-typography>`
+# `<clippy-token-sample-text>`
 
-Renders a sample of a typography token. Used to illustrate typography tokens and their concepts in the NL Design System.
+Renders a sample of a text token. Used to illustrate text tokens and their concepts in the NL Design System.
 
 ## Usage
 
 ```js
-import '@nl-design-system-community/clippy-components/clippy-token-sample-typography';
+import '@nl-design-system-community/clippy-components/clippy-token-sample-text';
 ```
 
 ```html
 <!-- standard -->
-<clippy-token-sample-typography></clippy-token-sample-typography>
+<clippy-token-sample-text></clippy-token-sample-text>
 ```
 
 ## Slots
