+++ packages/clippy-components/src/clippy-token-sample-text/styles.ts
@@ -12,7 +12,7 @@ export default css`
   }
 
   :host([truncate]) {
-    :where(.clippy-token-sample-typography__paragraph) {
+    :where(.clippy-token-sample-text__paragraph) {
       -webkit-box-orient: vertical;
       display: -webkit-box;
       -webkit-line-clamp: 1;
